requirements.txt
absl-py==1.1.0
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644569876605/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/ci/beautifulsoup4_1650293028159/work
bleach==5.0.0
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi @ file:///C:/ci_310/cffi_1643036627027/work
charset-normalizer==2.0.12
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.11.0
debugpy==1.6.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///C:/ci/entrypoints_1649926676279/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.15.3
Flask==2.1.2
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
grpcio==1.46.3
h5py==3.7.0
idna==3.3
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428305/work
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core @ file:///C:/ci/jupyter_core_1651638076360/work
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kafka==1.3.5
kafka-python==2.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune @ file:///C:/ci_310/mistune_1642084168466/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio @ file:///C:/ci/nest-asyncio_1649848132512/work
notebook==6.4.12
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
pywin32==304
pywinpty==2.0.5
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
spyder-kernels @ file:///Users/ktietz/demo/mc3/conda-bld/spyder-kernels_1629464163911/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado @ file:///C:/ci_310/tornado_1642093111997/work
traitlets==5.2.2.post1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.14.1
